requirements.txt
psycopg[binary]
openai
python-dotenv
